Service and folder renaming
