+++ src/main/java/uk/gov/hmcts/reform/roleassignment/domain/service/common/ParseRequestService.java
@@ -1,4 +1,6 @@
 package uk.gov.hmcts.reform.roleassignment.domain.service.common;
 
 public class ParseRequestService {
+    //1. Validate incoming data
+    //2. Mapping to model objects
 }
