+++ src/main/java/uk/gov/hmcts/reform/roleassignment/domain/service/common/PersistenceService.java
@@ -1,0 +1,9 @@
+package uk.gov.hmcts.reform.roleassignment.domain.service.common;
+
+public class PersistenceService {
+    //1. StoreRequest,
+    //2. StoreAssignment
+    //3. Update Request Status
+    //4. Update Assignment Status
+    //5. Make Assignment to Live
+}
