+++ src/main/java/uk/gov/hmcts/reform/roleassignment/domain/service/common/PrepareResponseService.java
@@ -1,4 +1,8 @@
 package uk.gov.hmcts.reform.roleassignment.domain.service.common;
 
 public class PrepareResponseService {
+    //1. prepare create 201 create assignment response
+    //2. prepare get single role assignment record
+    //3. prepare get multiple role assignment records
+    //4. Handle the mapping with HATEOUS if required.
 }
