+++ src/main/java/uk/gov/hmcts/reform/roleassignment/domain/service/common/RetrieveDataService.java
@@ -1,0 +1,8 @@
+package uk.gov.hmcts.reform.roleassignment.domain.service.common;
+
+public class RetrieveDataService {
+    //1. getting required case details(ccd data store)
+    //2. getting required role details(some static reference data??)
+    //3. getting required ticket details(Authorization table in JRD)
+    //4. getting some location reference data
+}
