+++ src/main/java/uk/gov/hmcts/reform/roleassignment/domain/service/common/ValidationModelService.java
@@ -1,0 +1,8 @@
+package uk.gov.hmcts.reform.roleassignment.domain.service.common;
+
+public class ValidationModelService {
+    //1. retrieve existingRoleAssignment records for Assignee
+    //2. retrieve existingRoleAssignment records for Requester
+    //3. retrieve AuthorisedRoleAssignment records for Requester/Assignee(??)
+    //Note: These are aggregation records from Assignment_history table.
+}
