--- src/main/java/uk/gov/hmcts/reform/roleassignment/domain/service/createresources/CreateResources.java
@@ -1,4 +1,0 @@
-package uk.gov.hmcts.reform.roleassignment.domain.service.createresources;
-
-public interface CreateResources {
-}
