--- src/main/java/uk/gov/hmcts/reform/roleassignment/domain/service/createresources/CreateRoleAssignments.java
@@ -1,7 +1,0 @@
-package uk.gov.hmcts.reform.roleassignment.domain.service.createresources;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class CreateRoleAssignments implements CreateResources{
-}
