+++ src/main/java/uk/gov/hmcts/reform/roleassignment/domain/service/createroles/CreateRoleAssignmentOrchestrator.java
@@ -1,0 +1,7 @@
+package uk.gov.hmcts.reform.roleassignment.domain.service.createroles;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class CreateRoleAssignmentOrchestrator {
+}
