--- src/main/java/uk/gov/hmcts/reform/roleassignment/domain/service/deleteresources/DeleteResources.java
@@ -1,4 +1,0 @@
-package uk.gov.hmcts.reform.roleassignment.domain.service.deleteresources;
-
-public interface DeleteResources {
-}
