--- src/main/java/uk/gov/hmcts/reform/roleassignment/domain/service/deleteresources/DeleteRoleAssignments.java
@@ -1,7 +1,0 @@
-package uk.gov.hmcts.reform.roleassignment.domain.service.deleteresources;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class DeleteRoleAssignments implements DeleteResources {
-}
