+++ src/main/java/uk/gov/hmcts/reform/roleassignment/domain/service/deleteroles/DeleteRoleAssignmentOrchestrator.java
@@ -1,0 +1,7 @@
+package uk.gov.hmcts.reform.roleassignment.domain.service.deleteroles;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class DeleteRoleAssignmentOrchestrator {
+}
