--- src/main/java/uk/gov/hmcts/reform/roleassignment/domain/service/getresource/RetrieveResource.java
@@ -1,7 +1,0 @@
-package uk.gov.hmcts.reform.roleassignment.domain.service.getresource;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public interface RetrieveResource {
-}
