--- src/main/java/uk/gov/hmcts/reform/roleassignment/domain/service/getresource/RetrieveRoleAssignment.java
@@ -1,4 +1,0 @@
-package uk.gov.hmcts.reform.roleassignment.domain.service.getresource;
-
-public class RetrieveRoleAssignment implements RetrieveResource {
-}
