+++ src/main/java/uk/gov/hmcts/reform/roleassignment/domain/service/getroles/RetrieveRoleAssignmentOrchestrator.java
@@ -1,0 +1,7 @@
+package uk.gov.hmcts.reform.roleassignment.domain.service.getroles;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class RetrieveRoleAssignmentOrchestrator {
+}
