--- src/main/java/uk/gov/hmcts/reform/roleassignment/domain/service/queryresources/QueryResources.java
@@ -1,7 +1,0 @@
-package uk.gov.hmcts.reform.roleassignment.domain.service.queryresources;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public interface QueryResources {
-}
