--- src/main/java/uk/gov/hmcts/reform/roleassignment/domain/service/queryresources/QueryRoleAssignments.java
@@ -1,4 +1,0 @@
-package uk.gov.hmcts.reform.roleassignment.domain.service.queryresources;
-
-public class QueryRoleAssignments implements QueryResources {
-}
