+++ src/main/java/uk/gov/hmcts/reform/roleassignment/domain/service/queryroles/QueryRoleAssignmentOrchestrator.java
@@ -1,0 +1,4 @@
+package uk.gov.hmcts.reform.roleassignment.domain.service.queryroles;
+
+public class QueryRoleAssignmentOrchestrator {
+}
