--- src/main/java/uk/gov/hmcts/reform/roleassignment/domain/service/validateresource/ValidateResourceOperation.java
@@ -1,7 +1,0 @@
-package uk.gov.hmcts.reform.roleassignment.domain.service.validateresource;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public interface ValidateResourceOperation {
-}
