--- src/main/java/uk/gov/hmcts/reform/roleassignment/domain/service/validateresource/ValidateRoleAssignmentOperation.java
@@ -1,4 +1,0 @@
-package uk.gov.hmcts.reform.roleassignment.domain.service.validateresource;
-
-public class ValidateRoleAssignmentOperation implements ValidateResourceOperation{
-}
