+++ src/main/java/uk/gov/hmcts/reform/roleassignment/domain/service/validateroles/RuleValidationOrchestrator.java
@@ -1,0 +1,4 @@
+package uk.gov.hmcts.reform.roleassignment.domain.service.validateroles;
+
+public class RuleValidationOrchestrator {
+}
