--- src/test/java/uk/gov/hmcts/reform/roleassignment/domain/service/createresources/CreateRoleAssignmentsTest.java
@@ -1,7 +1,0 @@
-package uk.gov.hmcts.reform.roleassignment.domain.service.createresources;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class CreateRoleAssignmentsTest {
-
-}
