--- src/test/java/uk/gov/hmcts/reform/roleassignment/domain/service/deleteresources/DeleteRoleAssignmentsTest.java
@@ -1,7 +1,0 @@
-package uk.gov.hmcts.reform.roleassignment.domain.service.deleteresources;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class DeleteRoleAssignmentsTest {
-
-}
