--- src/test/java/uk/gov/hmcts/reform/roleassignment/domain/service/getresource/RetrieveRoleAssignmentTest.java
@@ -1,7 +1,0 @@
-package uk.gov.hmcts.reform.roleassignment.domain.service.getresource;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class RetrieveRoleAssignmentTest {
-
-}
