--- src/test/java/uk/gov/hmcts/reform/roleassignment/domain/service/queryresources/QueryRoleAssignmentsTest.java
@@ -1,7 +1,0 @@
-package uk.gov.hmcts.reform.roleassignment.domain.service.queryresources;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class QueryRoleAssignmentsTest {
-
-}
