+++ src/test/java/uk/gov/hmcts/reform/roleassignment/domain/service/queryroles/QueryRoleAssignmentsTest.java
@@ -1,0 +1,5 @@
+package uk.gov.hmcts.reform.roleassignment.domain.service.queryroles;
+
+class QueryRoleAssignmentsTest {
+
+}
