--- src/test/java/uk/gov/hmcts/reform/roleassignment/domain/service/validateresource/ValidateRoleAssignmentOperationTest.java
@@ -1,7 +1,0 @@
-package uk.gov.hmcts.reform.roleassignment.domain.service.validateresource;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class ValidateRoleAssignmentOperationTest {
-
-}
